Added a new scene for interest testing. I think splitting scenes up for tests will be better for commits going forward.
Added an interest value to points and agents that will determine if they are "interested"
Attempted to have the agents face the interest. (Still WIP)

Added a sprint function to the camera. Holding lshift will increase speed by 5 times.

Messed with character controller script to control an individual agent. not a priority, doesn't work.